Follow activate order for after_build extensions
The registering of the after_build hook was delayed by the execution of the after_configuration hook. That way it is not possible to define the order of execution of after_build extensions that normally would follow the activate order in the config.rb of a Middleman project.

This is a issue when syncing to S3 in the build phase. You want to smush BEFORE syncing.