Remove Middleman::Smusher::VERSION
Since the separate file was defining a Middleman::Smusher module, while in fact, Middleman::Smusher is a subclass of Middleman::Extension. This caused conflicts, when loading the extension.

Removing the separate file and adding the version number directly to the gemspec was the simplest solution. Changing the declaration within version.rb to the proper class definition would defeat the idea of a simple, easy-to-load version file, which can be used from the gemspec without further prerequisites.